Move IndesUtils to o.a.l.luke.models package
