+++ src/main/java/org/apache/lucene/luke/app/DirectoryHandler.java
@@ -19,7 +19,7 @@ package org.apache.lucene.luke.app;
 
 import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 import org.apache.lucene.luke.models.LukeException;
-import org.apache.lucene.luke.util.IndexUtils;
+import org.apache.lucene.luke.models.util.IndexUtils;
 import org.apache.lucene.store.Directory;
 
 import java.io.IOException;
