+++ src/main/java/org/apache/lucene/luke/app/IndexHandler.java
@@ -20,7 +20,7 @@ package org.apache.lucene.luke.app;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 import org.apache.lucene.luke.models.LukeException;
-import org.apache.lucene.luke.util.IndexUtils;
+import org.apache.lucene.luke.models.util.IndexUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
